requirements.txt
discord.py[voice]>=2.4.0
discord-ext-voice-recv>=0.5.2
openai>=1.57.0
httpx>=0.27.0
python-dotenv>=1.0.1
pytest>=8.3.0
pytest-asyncio>=0.24.0
